+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PR template
 
+### Changed
+
+- FIX typo in marketing cookie text
+
 ## [1.5.2] - 2024-07-08
 
 ### Changed